requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-jobspy
pandas
apscheduler
tzdata
asyncpg
greenlet
